backend/requirements.txt
asgiref==3.5.2
Django==4.0.4
djoser==2.1.0
django-cors-headers==3.12.0
djangorestframework==3.13.1
Pillow==9.1.0
pytils==0.4.1
pytz==2022.1
slugify==0.0.1
sqlparse==0.4.2
celery==5.2.7
redis==4.3.4
python-telegram-bot==13.13
requests==2.27.1